Mex compiles and runs
I put the EIGEN_MAKE_ALIGNED_OPERATOR_NEW directive in every header file to hopefully fix some of the crashes. It didn't seem to work. Setting the number of threads to 1 did. It seems theres some memory/heap allocation issues between MEX and eigen that get revealed when we are multithreading. The directive was supposed to address that but the polymorphism in the FEM_Simulator for TimeIntegration may be adding another layer of difficulty